+++ CHANGELOG.md
@@ -2,3 +2,12 @@
 
 * Basic backdrop implementation.
 * BackdropScaffold widget implemented.
+
+## [0.1.0] - 28 July 2018.
+
+* opacity fixes in BackdropScaffold.
+* headerHeight option added in BackdropScaffold.
+* frontLayerBorderRadius option added in BackdropScaffold
+* refactor: body -> frontLayer
+* refactor: backpanel -> backLayer
+* BackdropToggleButton widget implemented.